7/25일 회의 수정건 중간 commit
- [x]  partyFragment, RestaurantParty 에 있는 더미 버튼 삭제
- [ ]  정원 선택된 글씨 조금 커지게
- [x]  테스트 공지, 이름 textview 왼쪽 좀 띄워주기
- [x]  파티 생성 이후 오류 // 생성이 되긴 함
- [x]  파티 채팅 시 (텍스트, 이미지) 종료현상 // 채팅 전달은 잘 되는 듯
- [x]  파티리스트화면에서 리사이클러뷰가 잘려있음
- [x]  파티 참여에 중복으로 참여 가능 → 주석처리 해제할 것
- [x]  네비게이션바의 파티 프래그먼트로 전환 시 종료
- [x]  파티 참여 시 정원 꽉 차면  못 들어감 → 안내메시지 필요!!
- [x]  파티정보수정은 파티장만 가능한지 확인
- [x]  파티생성화면 그룹명 → 파티
- [x]  파티생성화면의 취소버튼 → 삭제( 취소버튼에 기능만들어서 다시 만들까요?)
- [x]  파티장 표시 안 뜸  → 파이어베이스에 user6(곰)가 안들어가 있었음 추가했더니 정상작동/ 지속적으로 확인해보고 안되면 로직변경 해봄

+++ app/src/main/java/com/bapool/bapool/retrofit/data/Party.kt
@@ -94,7 +94,7 @@ data class FirebasePartyInfo
     val endDate: String = "",
     val hashTag: List<Int> = listOf(),
     val restaurantName: String = "",
-    val groupLeaderId: Int = 0,
+    val groupLeaderId: Long = 0,
     val siteUrls: String = "",
 ): Serializable
